Move gRPC client to cmd/client
